Fixing collision detection algorithm and adding contain check.
Algorithm used requires origin to be at the center of the box,
which is not the case here. Origin is at bottom left corner.

Adding method to check if a bounding box is inside another.
Also renaming both collision methods to start with 'is...'.

New constructor for Rect, to be created by passing
it's coordinates and sizes.

Signed-off-by: rondao <rafael.rondao@gmail.com>